HIVE-7527 - Support order by and sort by on Spark (Rui Li via Brock) [Spark Branch]
git-svn-id: https://svn.apache.org/repos/asf/hive/branches/spark@1615864 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/plan/SparkEdgeProperty.java
@@ -33,8 +33,8 @@ public class SparkEdgeProperty {
     this.edgeType = edgeType;
     this.numPartitions = numPartitions;
   }
-  
-  public SparkEdgeProperty(int edgeType) {
+
+  public SparkEdgeProperty(long edgeType) {
     this.edgeType = edgeType;
   }
